Move ManagerSettings into settings header so the new storage header doesn't have to include the whole runtime header
