+++ lib/config/sqlite_storage.cpp
@@ -7,7 +7,9 @@
 #include <everest/database/exceptions.hpp>
 #include <everest/database/sqlite/schema_updater.hpp>
 #include <everest/logging.hpp>
+
 #include <utils/config/storage_sqlite.hpp>
+#include <utils/date.hpp>
 
 using json = nlohmann::json;
 using namespace everest::db;
